backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
google-generativeai>=0.4.1,<0.5.0
python-dotenv==1.0.0
requests==2.31.0
pymongo==4.6.1
python-multipart==0.0.9
